refactoring/Checkin: extract functions from Checkin/Emotions/RecordProgress/User-Providers
